kernel: sched: fair: Fix u64 load_sum truncation in max_t(u32, ...)
load_sum is u64 but max_t(u32, ...) silently truncates both operands to
32 bits before comparison. On busy systems where load_sum exceeds 2^32,
this corrupts PELT load tracking. Use max_t(u64, ...) instead.

+++ kernel/sched/fair.c
@@ -3107,7 +3107,7 @@ dequeue_load_avg(struct cfs_rq *cfs_rq, struct sched_entity *se)
 	sub_positive(&cfs_rq->avg.load_avg, se->avg.load_avg);
 	sub_positive(&cfs_rq->avg.load_sum, se_weight(se) * se->avg.load_sum);
 	/* See update_cfs_rq_load_avg() */
-	cfs_rq->avg.load_sum = max_t(u32, cfs_rq->avg.load_sum,
+	cfs_rq->avg.load_sum = max_t(u64, cfs_rq->avg.load_sum,
 					  cfs_rq->avg.load_avg * PELT_MIN_DIVIDER);
 }
 #else
@@ -3799,7 +3799,7 @@ update_tg_cfs_load(struct cfs_rq *cfs_rq, struct sched_entity *se, struct cfs_rq
 	add_positive(&cfs_rq->avg.load_avg, delta_avg);
 	add_positive(&cfs_rq->avg.load_sum, delta_sum);
 	/* See update_cfs_rq_load_avg() */
-	cfs_rq->avg.load_sum = max_t(u32, cfs_rq->avg.load_sum,
+	cfs_rq->avg.load_sum = max_t(u64, cfs_rq->avg.load_sum,
 					  cfs_rq->avg.load_avg * PELT_MIN_DIVIDER);
 }
